requirements.txt
pinecone-client==5.0.1
langchain_core==0.2.34
pdf2image==1.17.0
pytesseract==0.3.13
langchain==0.2.14
langchain_openai==0.1.22
langchain_pinecone==0.1.3
gunicorn==23.0.0
python-dotenv==1.0.1
Flask==3.0.3